+++ Tests/Unit/Public/Protect-GuestConfigurationPackage.Tests.ps1
@@ -1,15 +1,10 @@
 BeforeDiscovery {
 
-    $script:projectPath = "$PSScriptRoot/../../.." | Convert-Path
-    $script:projectName = Get-SamplerProjectName -BuildRoot $script:projectPath
+    $projectPath = "$PSScriptRoot/../../.." | Convert-Path
+    $projectName = Get-SamplerProjectName -BuildRoot $projectPath
 
-    Get-Module $script:projectName | Remove-Module -Force -ErrorAction SilentlyContinue
-    $script:importedModule = Import-Module $script:projectName -Force -PassThru -ErrorAction 'Stop'
-
-
-
-    $IsNotAzureDevOps = [string]::IsNullOrEmpty($env:ADO)
-    $IsNotWindowsAndIsAzureDevOps = ($IsLinux -or $IsMacOS) -AND $env:ADO
+    Get-Module $projectName | Remove-Module -Force -ErrorAction SilentlyContinue
+    $importedModule = Import-Module $projectName -Force -PassThru -ErrorAction 'Stop'
 
     $IsRunningAsAdmin = $false
 
@@ -20,10 +15,14 @@ BeforeDiscovery {
     }
 }
 
-Context 'Protect-GuestConfigurationPackage' {
+Describe 'Protect-GuestConfigurationPackage' -ForEach @{
+    ProjectPath    = $projectPath
+    projectName    = $projectName
+    importedModule = $importedModule
+} {
 
     BeforeAll {
-        $testHelpersPath = Join-Path -Path $script:projectPath -ChildPath 'Tests/helpers'
+        $testHelpersPath = Join-Path -Path $projectPath -ChildPath 'Tests/helpers'
         Get-ChildItem -Path $testHelpersPath -Filter *.ps1 -Recurse | Foreach-Object {
             # dot sourcing the helpers files
             Write-Host -Object "`tImporting helper file $($_.Name)"