+++ source/Modules/DscOperations/Public/Start-DscConfiguration.ps1
@@ -1,0 +1,33 @@
+<#
+    .SYNOPSIS
+        Start DSC configuration.
+
+    .Parameter ConfigurationName
+        Configuration name.
+
+    .Example
+        Start-DscConfiguration -ConfigurationName WindowsTLS
+#>
+function Start-DscConfiguration
+{
+    [CmdletBinding()]
+    param (
+        [Parameter(Mandatory = $true)]
+        [ValidateNotNullOrEmpty()]
+        [System.String]
+        $ConfigurationName
+    )
+
+    $job_id = [guid]::NewGuid().Guid
+    $gcBinPath = Get-GuestConfigBinaryPath
+    $dsclibPath = $(Get-DscLibPath) -replace  '[""\\]','\$&'
+
+    if (-not ([System.Management.Automation.PSTypeName]'GuestConfig.DscOperations').Type)
+    {
+        $addTypeScript = $ExecuteDscOperationsScript -f $dsclibPath
+        Add-Type -TypeDefinition $addTypeScript -ReferencedAssemblies 'System.Management.Automation','System.Console','System.Collections'
+    }
+
+    $dscOperation = [GuestConfig.DscOperations]::New()
+    $dscOperation.StartDscConfiguration($PSCmdlet, $job_id, $ConfigurationName, $gcBinPath, $True, $True)
+}